+++ lord/settings.py
@@ -1,0 +1,5 @@
+import os
+
+LOCAL_HOME = os.environ.get('HOME')
+LOCAL_DATA = os.path(['.local','share','lord']
+LOCAL_SAVED = ['saved']