+++ tests/test_concretas.py
@@ -1,0 +1,7 @@
+from lord.concretas import CartaConcreta, JogoConcreto
+
+def test_atualizar():
+    oi =      'dsga'
+#    jogo = JogoConcreto()
+#    carta = CartaConcreta("Carta")
+#    jogo.anexar(carta)