+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Não Publicado]
+- Implementar o padrão Observador para a carta de objetivo saber que está defendida/guardada.
 
 ## 0.2.0 [2023-06-10]
 ### Adicionado